pyproject.toml
[tool.poetry]
name = "trading-dashboard"
version = "0.1.0"
description = "Stock analysis dashboard using streamlit"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.41.1"
playwright = "^1.49.1"
supabase = "^2.10.0"
plotly = "^5.24.1"
yfinance = "^0.2.50"
python-telegram-bot = "^21.9"
colorama = "^0.4.6"
email-validator = "^2.2.0"
inquirer = "^3.4.0"
prompt-toolkit = "^3.0.48"
loguru = "^0.7.3"
pydantic = "^2.10.4"
jinja2 = "^3.1.5"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
pyright = "^1.1.391"
watchdog = "^6.0.0"
ipdb = "^0.13.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
